requirements.txt
python-telegram-bot==21.7
gspread==6.1.4
google-auth==2.36.0
python-dotenv==1.0.1
